tweak

+++ include/simdjson/generic/ondemand/parser-inl.h
@@ -44,7 +44,7 @@ simdjson_warn_unused simdjson_inline error_code parser::allocate(size_t new_capa
 }
 #if SIMDJSON_DEVELOPMENT_CHECKS
 simdjson_inline simdjson_warn_unused bool parser::string_buffer_overflow(const uint8_t *string_buf_loc) const noexcept {
-  return (string_buf_loc < string_buf.get()) || (string_buf_loc - string_buf.get() >= capacity());
+  return (string_buf_loc < string_buf.get()) || (size_t(string_buf_loc - string_buf.get()) >= capacity());
 }
 #endif
